requirements.txt
streamlit>=1.45.0
rembg>=2.0.67
pillow>=11.0.0
onnxruntime>=1.20.0
requests
opencv-python
numpy
pytesseract
reportlab